backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.8.2
httpx==0.26.0
faiss-cpu==1.13.0
numpy==1.26.4
langchain==0.2.16
langgraph==0.1.8
langchain-community==0.2.6
python-multipart==0.0.9
